backend/requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic
psycopg2-binary
pydantic[email]
pydantic==2.9.1
pydantic_core==2.23.3
python-multipart
reportlab==4.2.2
pyjwt==2.9.0
passlib==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1